Relanding the following:
When restoring a session, the browser is shown after a tab is selected causing
the focus to be on the location bar (not tab contents for web pages).
This CL ensures the tab decides where the focus should be after the browser is
shown.

BUG=6000
TEST=See bug
R=sky
Review URL: http://codereview.chromium.org/28156

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@10488 0039d316-1c4b-4281-b951-d872f2087c98