Fixing crash in StartFinding (2nd attempt).
The first attempt to fix this was unsuccessful and crashes showed
up again during the automated ui testing. I still can't reproduce
locally, but when I look again at the crash dump I think that the
fact that it looked like the FindBarView was destroyed is misleading
and the real issue is the same as with the crash in StopFinding
(that web_contents has been destroyed moments before we receive the
request).

I am therefore removing the code I added earlier and replacing it
with a null check for web_contents.

TEST=Covered by automated ui tests.
BUG=8048

Review URL: http://codereview.chromium.org/27219

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@10489 0039d316-1c4b-4281-b951-d872f2087c98