Make worker constructor return error if it is enabled by the flag.
This is to move corresponding files out of http://codereview.chromium.org/48106/ so that this can be checked in before the change to enable worker. The reason for doing this is to avoid build break since the already checked-in third_party/WebKit/WebKit/chromium/src/WebKit.cpp  has the ifdef guarded code depend on this.
Review URL: http://codereview.chromium.org/42313

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@11952 0039d316-1c4b-4281-b951-d872f2087c98