Relanding since it was breaking the Release build (because of missing includes).
Moving the mouse-wheel message rerouting code out of the focus manager.
This is part of the effort of making the focus manager not window specific.

BUG=None
TEST=Make sure mouse-wheel scrolls the Chrome window under the cursor (try with
a Chrome window, the option window, the task manager table) when it is active
and when it is inactive. Also try using the mouse-wheel on a plugin with a
scroll-view (for example
http://java.sun.com/products/plugin/1.4/demos/plugin/jfc/SwingSet2/SwingSet2.html)
when the browser window is not active, the plugin should scroll.
TBR=ben

Review URL: http://codereview.chromium.org/42624

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@12517 0039d316-1c4b-4281-b951-d872f2087c98