A quick fix for Issue 11683.
I noticed this issue is caused by the IMF_AUTOKEYBOARD option of a RichEdit control, an option that allows a RichEdit control to automatically change the input language with a selection. (An omnibox changes a selection to display a styled text.) A RichEdit control automatically turns on the IMF_AUTOKEYBOARD option when we input an RTL character and start changing the input language with a selection. So, this change checks the IMF_AUTOKEYBOARD option and disable it before decorating a text to prevent it from changing the input language while decorating a text.

BUG=11683 "RTL: IME changes when writing a Hebrew character inside an English string in the omnibox"
TEST=Change the input language to English, type "abcd" in an omnibox, change the input language to Hebrew, type an 'a' key, and verify the input language is Hebrew.
Review URL: http://codereview.chromium.org/113394

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@16154 0039d316-1c4b-4281-b951-d872f2087c98