Fix crash 12519: in views::DialogClientView::AcceptWindow()
We don't have repro steps for this crash so this
is a blind fix.

But looking at the function that crashes and the
history of this file I see:

Ben's r11070 and r11047: Which add this check.
Ben's r11132: Which removes this check (according
to the comments it is a test to see if we can get
away with it).

It looks to me like the check would prevent the
crash and since this is still crashing, I am
proposing we re-add it.

BUG=12519
TEST=We'll need to verify this by checking the
crash dumps sent from the field.

Review URL: http://codereview.chromium.org/119183

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@17651 0039d316-1c4b-4281-b951-d872f2087c98