Initial download shelf on OS X.
This has lots of missing stuff (e.g. a custom download item view that shows
download progress, the popup is the same for in-progress and completed
downloads, no animation, everything looks ugly, the info bubble overlaps the
shelf when it's visible, no "open download manager page" link, etc), but the
basic functionality is hooked up: The shelf appears when files are downloaded,
and something ugly is added to the shelf for each download. The popup's "Reveral
in Finder" even works. The shelf is per-window as it should be.

BUG=12500
TEST=Download something and check the shelf appears. Click the close button and
make sure it disappears again.

Review URL: http://codereview.chromium.org/93129


git-svn-id: svn://svn.chromium.org/chrome/trunk/src@18757 0039d316-1c4b-4281-b951-d872f2087c98